mcp/requirements.txt
flask>=2.0.0
requests>=2.25.0
websocket-client>=1.0.0